Several changes for the Alpaka version
1. HostDevDef.h

Commented out the ALPAKA block. It turns out it is not used at all for the
CUDA backend, as for other backends (e.g. HIP) it needs to be moved up
in the if-else chain, and un-commented

2. FastCaloGpu/src/CMakeLists.txt

Don't attempt to configure CUDA if alpaka is used with non-CUDA backend

3. GeoLoadGpu_al.cxx

Fixed compiler warnings related to format mismatch in printf()